+++ examples/meshbatch.rs
@@ -39,8 +39,8 @@ impl MainState {
 
         // Generate enough instances to fill the entire screen
         let size = ctx.gfx.drawable_size();
-        let items_x = (size.0 / 16.0) as u32;
-        let items_y = (size.1 / 16.0) as u32;
+        let items_x = (size.0 / 16.0) as usize;
+        let items_y = (size.1 / 16.0) as usize;
         let mut mesh_batch = graphics::InstanceArray::new(ctx, None);
         mesh_batch.resize(ctx, items_x * items_y);
